requirements.txt
rembg[cpu]==2.0.59
onnxruntime==1.18.0
Pillow==10.4.0
google-api-python-client==2.140.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
tenacity==8.5.0
python-dotenv==1.0.1
psutil==6.1.0
